requirements.txt
taichi~=1.7.4
numpy~=2.4.2
pygame
pyopengl
pyrr